+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Updated
 1. Reworked event handling to use connected UDP socket to send events.
 2. Updated to Go 1.23.
+3. Reworked `put-card` handler to return false if card _from_ or _to_ date is a zero value.
 
 
 ## [0.8.8](https://github.com/uhppoted/uhppote-simulator/releases/tag/v0.8.8) - 2024-03-27