+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Updated
 1. Updated to Go 1.24.
+2. Fixed incoming requests not printed to console bug.
 
 
 ## [0.8.10](https://github.com/uhppoted/uhppote-simulator/releases/tag/v0.8.10) - 2025-01-30